requirements.txt
# Telegram Trading Bot Dependencies

# Telegram User Client
telethon>=1.34.0

# Blockchain - EVM
web3>=6.15.0

# Blockchain - Solana
solana>=0.32.0
solders>=0.21.0

# HTTP & Async
aiohttp>=3.9.0
httpx>=0.27.0

# Database
aiosqlite>=0.19.0

# Configuration
python-dotenv>=1.0.0

# Security
cryptography>=42.0.0

# Utilities
pydantic>=2.5.0
